// src/modules/operadores/interfaces/operador.ts
export interface Operador {
  id: number;
  nombre: string;
  paterno: string;
  materno: string;
  fecha_nacimiento: string;
  fecha_ingreso: string;
  curp: string;
  rfc: string;
  codigo_postal: string;
  colonia: string;
  municipio: string;
  estado: string;
  pais: string;
  calle: string;
  numero_exterior: string;
  numero_interior: string;
  licencia: string;
  vencimiento_licencia: string;
  estatus: boolean;
  gafete: string;
}